+++ source-code/Client/frontendandroid/app/src/main/java/com/iasdietideals24/dietideals24/utilities/classes/data/Asta.kt
@@ -1,5 +1,6 @@
 package com.iasdietideals24.dietideals24.utilities.classes.data
 
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalTime
 
@@ -9,7 +10,7 @@ data class Asta(
     private val _tipo: String = "",
     private val _dataFine: LocalDate = LocalDate.MIN,
     private val _oraFine: LocalTime? = LocalTime.MIN,
-    private val _prezzo: Double = 0.0,
+    private val _prezzo: BigDecimal = BigDecimal(0.0),
     private val _immagine: ByteArray = ByteArray(0),
     private val _nome: String = "",
     private val _categoria: String = "",
