+++ source-code/Client/frontendandroid/app/src/main/java/com/iasdietideals24/dietideals24/utilities/classes/data/OffertaRicevuta.kt
@@ -1,5 +1,6 @@
 package com.iasdietideals24.dietideals24.utilities.classes.data
 
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalTime
 
@@ -9,7 +10,7 @@ data class OffertaRicevuta(
     val idOfferente: Long = 0L,
     val nomeOfferente: String = "",
     val immagineOfferente: ByteArray = ByteArray(0),
-    val offerta: Double = 0.0,
+    val offerta: BigDecimal = BigDecimal(0.0),
     val dataInvio: LocalDate = LocalDate.MIN,
     val oraInvio: LocalTime = LocalTime.MIN,
     val accettata: Boolean? = null
